Code Cleanup

+++ src/main/java/net/minecraft/client/model/TexturedQuad.java
@@ -47,9 +47,9 @@ public class TexturedQuad
         Vector3D vector3D = this.vertexPositions[1].vector3D.subtractReverse(this.vertexPositions[0].vector3D);
         Vector3D vec31D = this.vertexPositions[1].vector3D.subtractReverse(this.vertexPositions[2].vector3D);
         Vector3D vec32D = vec31D.crossProduct(vector3D).normalize();
-        float f = (float) vec32D.x;
-        float f1 = (float) vec32D.y;
-        float f2 = (float) vec32D.z;
+        float f = (float) vec32D.x();
+        float f1 = (float) vec32D.y();
+        float f2 = (float) vec32D.z();
 
         if (this.invertNormal)
         {
@@ -70,7 +70,7 @@ public class TexturedQuad
         for (int i = 0; i < 4; ++i)
         {
             PositionTextureVertex positiontexturevertex = this.vertexPositions[i];
-            renderer.pos(positiontexturevertex.vector3D.x * (double)scale, positiontexturevertex.vector3D.y * (double)scale, positiontexturevertex.vector3D.z * (double)scale).tex((double)positiontexturevertex.texturePositionX, (double)positiontexturevertex.texturePositionY).normal(f, f1, f2).endVertex();
+            renderer.pos(positiontexturevertex.vector3D.x() * (double)scale, positiontexturevertex.vector3D.y() * (double)scale, positiontexturevertex.vector3D.z() * (double)scale).tex(positiontexturevertex.texturePositionX, positiontexturevertex.texturePositionY).normal(f, f1, f2).endVertex();
         }
 
         Tessellator.getInstance().draw();
